Extrema (min es max)

+++ src/SigStat.Common/Features/Features.cs
@@ -16,7 +16,7 @@ namespace SigStat.Common
         public static readonly FeatureDescriptor<int> Dpi = new FeatureDescriptor<int>("DPI", "Dpi");
         public static readonly FeatureDescriptor<List<double>> X = new FeatureDescriptor<List<double>>("X(t)", "X");
         public static readonly FeatureDescriptor<List<double>> Y = new FeatureDescriptor<List<double>>("Y(t)", "Y");
-        public static readonly FeatureDescriptor<List<int>> T = new FeatureDescriptor<List<int>>("t", "Svc2004.t");
+        public static readonly FeatureDescriptor<List<double>> T = new FeatureDescriptor<List<double>>("t", "Svc2004.t");
         public static readonly FeatureDescriptor<List<int>> Button = new FeatureDescriptor<List<int>>("Button(t)", "Button");
         public static readonly FeatureDescriptor<List<int>> Azimuth = new FeatureDescriptor<List<int>>("Azimuth(t)", "Azimuth");
         public static readonly FeatureDescriptor<List<int>> Altitude = new FeatureDescriptor<List<int>>("Altitude(t)", "Altitude");
